style: add custom style to toolbar & refactor components in directories
