+++ src/components/App.js
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import store from '../stores/MainStore'
-import Canvas from './Canvas'
-import { Toolbar } from './Toolbar'
+import Canvas from './Canvas/Canvas'
+import { Toolbar } from './Toolbar/Toolbar'
 import { observer } from 'mobx-react'
 
 function App () {
