+++ src/components/Box/Box.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { observer } from 'mobx-react'
 import BoxDraggable from './BoxDraggable'
+import './Box.css'
 
 function Box(props) {
   return (
