+++ src/components/Canvas/Canvas.css
@@ -1,0 +1,6 @@
+.canvas {
+    width: 1200px;
+    height: 675px;
+    background-color: aliceblue;
+}
+
