+++ src/components/Toolbar/Toolbar.css
@@ -1,0 +1,35 @@
+
+.toolbar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    display: flex;
+    padding: 1rem 0.5rem;
+    gap: 1rem;
+    border: 2px dashed #b4b4b4;
+    border-radius: 1rem;
+    margin: 0.5rem;
+    align-items: center;
+}
+
+.toolbar span {
+    color: white;
+}
+
+.toolbar button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0.5rem 1rem;
+    border: 1px solid #b4b4b4;
+    border-radius: 0.5rem;
+    background-color: #282c34;
+    color: white;
+    font-weight: 600;
+    text-transform: uppercase;
+    cursor: pointer;
+}
+
+.toolbar button:hover {
+    background-color: #b4b4b4;
+}
