+++ src/components/Toolbar/Toolbar.js
@@ -1,8 +1,9 @@
 import React from 'react'
-import store from '../stores/MainStore'
+import store from '../../stores/MainStore'
 import { observer } from 'mobx-react'
 import { toast, ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
+import './Toolbar.css'
 
 export const Toolbar = observer(() => {
   const handleAddButton = () => {
