+++ src/main.css
@@ -18,26 +18,5 @@ html,
   align-items: center;
   justify-content: center;
 }
-.canva {
-  width: 1200px;
-  height: 675px;
-  background-color: aliceblue;
-}
 
-.box {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
 
-.toolbar {
-  position: fixed;
-  top: 0;
-  left: 0;
-  display: flex;
-}
-
-.toolbar span {
-  color: white;
-}
